PR #396: Marker commit for authorship info
This entire PR (of which this commit is the capstone) was co-authored by @ZeldaZach.
In the process of editing his code the original authorship information was lost,
and it ended up being a mix of our code that made it into the final PR.

As a way to ensure he is still at least somewhat credited for this commit series,
this empty commit lists him as author so he is counted as a contributor in Github stats.