thrimbletrimmer: Show any chat we have even if no end time set
This does not auto refresh, because that wouldn't be useful,
because chat doesn't reach disk until ~1min after being recorded. You can't see it live.

It is still useful to eg. see the first 58 minutes of chat if you load a 1h video
but the padding puts your end time "in the future".