thrimbletrimmer: hacky way to avoid loading large videos
Partially implements #382, by adding a query param `load_video=false` which makes the
main video not load. This is only enabled explicitly because this leaves the UI in a quite broken state.
You cannot submit changes, see time ranges or transitions, or create thumbnails.

About all you CAN do is use the manual link and reset row buttons.
This suffices for the primary use-case, which is setting the manual link for Supercuts
that span several days.