Parse chapter markers out of the description
This uses a new format in the description, meaning the description can no longer be sent to YouTube without being loaded into the editor and explicitly submitted. (Not that there was really another way to cause a video to become sent to YouTube before, but now this step is required.)
The new format for chapter markers, for each line, when saving as draft, is:
index;wubloader_time - description

When submitting a video, times will be validated according to the restrictions imposed by YouTube that we know about (based on the old Thrimbletrimmer) and converted to video play times with second granularity.